wip logbuff class
