+++ src/classes/log.js
@@ -1,5 +1,6 @@
 import { v4 as uuid } from 'uuid';
 
+// Used for logging.
 export default class Log {
   constructor(content) {
     this.uuid = uuid();
